# artifacts/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PSF Lab local API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/traces": {
      "get": {
        "summary": "Traces",
        "operationId": "traces_api_traces_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Upload",
        "operationId": "upload_api_traces_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_traces_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/traces/{id_}": {
      "get": {
        "summary": "Metadata",
        "operationId": "metadata_api_traces__id___get",
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/traces/{id_}/events": {
      "post": {
        "summary": "Events",
        "operationId": "events_api_traces__id___events_post",
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/traces/{id_}/view": {
      "post": {
        "summary": "View",
        "operationId": "view_api_traces__id___view_post",
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/traces/{id_}/export": {
      "post": {
        "summary": "Export",
        "operationId": "export_api_traces__id___export_post",
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs": {
      "get": {
        "summary": "Run List",
        "operationId": "run_list_api_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/runs/{id_}/psf": {
      "get": {
        "summary": "Run Psf",
        "operationId": "run_psf_api_runs__id___psf_get",
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/comparisons": {
      "post": {
        "summary": "Compare",
        "operationId": "compare_api_comparisons_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_api_traces_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_api_traces_post"
      },
      "CompareRequest": {
        "properties": {
          "pair_id": {
            "type": "string",
            "title": "Pair Id"
          },
          "run_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Run Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pair_id",
          "run_ids"
        ],
        "title": "CompareRequest"
      },
      "ExportRequest": {
        "properties": {
          "filters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filters"
          },
          "sort": {
            "items": {},
            "type": "array",
            "title": "Sort"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "events"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ExportRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "QueryRequest": {
        "properties": {
          "filters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filters"
          },
          "sort": {
            "items": {},
            "type": "array",
            "title": "Sort"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "default": 0
          },
          "limit": {
            "type": "integer",
            "maximum": 2000.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 200
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "QueryRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ViewRequest": {
        "properties": {
          "filters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filters"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ViewRequest"
      }
    }
  }
}